requirements.txt
fastapi
uvicorn
python-dotenv
python-telegram-bot==20.3
gunicorn
apscheduler
